Added test for saving a Proxy

+++ src/test/java/org/doogie/liquido/test/TestFixtures.java
@@ -9,7 +9,7 @@ public class TestFixtures {
   public static final String USER1_PWD   = "dummyPasswordHash";
   public static final String USER4_EMAIL = "testuser4@liquido.de";
   public static final String AREA1_TITLE = "Area 1";
-  public static final int USER4_NUM_VOTES = 5;
+  public static final long   USER4_NUM_VOTES = 5;     // testuser4@liquido.de  has 5 votes (including his own) due to (transitive) delegations
 
 
 }
